libs/hardware_interface/requirements.txt
pyflakes==3.4.0
pyserial==3.5
smbus2==0.5.0
